+++ .eslintrc.js
@@ -100,7 +100,8 @@ module.exports = {
   },
   "plugins": [
     "@typescript-eslint",
-    "cli-microsoft365"
+    "cli-microsoft365",
+    "mocha"
   ],
   "ignorePatterns": [
     "**/pcf-init/assets/**",
@@ -186,7 +187,8 @@ module.exports = {
         ]
       }
     ],
-    "@typescript-eslint/explicit-function-return-type": ["error", { "allowExpressions": true }]
+    "@typescript-eslint/explicit-function-return-type": ["error", { "allowExpressions": true }],
+    "mocha/no-identical-title": "error"
   },
   "overrides": [
     {