+++ src/cli/Cli.ts
@@ -8,7 +8,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { Logger } from './Logger';
 import appInsights from '../appInsights';
-import Command, { CommandArgs, CommandError } from '../Command';
+import Command, { CommandArgs, CommandError, CommandTypes } from '../Command';
 import config from '../config';
 import GlobalOptions from '../GlobalOptions';
 import request from '../request';
@@ -18,6 +18,7 @@ import { fsUtil } from '../utils/fsUtil';
 import { md } from '../utils/md';
 import { CommandInfo } from './CommandInfo';
 import { CommandOptionInfo } from './CommandOptionInfo';
+import { validation } from '../utils/validation';
 const packageJSON = require('../../package.json');
 
 export interface CommandOutput {
@@ -38,6 +39,8 @@ export class Cli {
   private optionsFromArgs: { options: minimist.ParsedArgs } | undefined;
   public commandsFolder: string = '';
   private static instance: Cli;
+  private static defaultHelpMode = 'full';
+  public static helpModes: string[] = ['options', 'examples', 'remarks', 'response', 'full'];
 
   private _config: Configstore | undefined;
   public get config(): Configstore {
@@ -107,9 +110,17 @@ export class Cli {
       // we have found a command to execute. Parse args again taking into
       // account short and long options, option types and whether the command
       // supports known and unknown options or not
-      this.optionsFromArgs = {
-        options: this.getCommandOptionsFromArgs(rawArgs, this.commandToExecute)
-      };
+
+      try {
+        this.optionsFromArgs = {
+          options: this.getCommandOptionsFromArgs(rawArgs, this.commandToExecute)
+        };
+      }
+      catch (e: any) {
+        const optionsWithoutShorts = Cli.removeShortOptions({ options: parsedArgs });
+
+        return this.closeWithError(e.message, optionsWithoutShorts, false);
+      }
     }
     else {
       this.optionsFromArgs = {
@@ -123,7 +134,7 @@ export class Cli {
       showHelp ||
       parsedArgs.h ||
       parsedArgs.help) {
-      this.printHelp();
+      this.printHelp(this.getHelpMode(parsedArgs));
       return Promise.resolve();
     }
 
@@ -248,6 +259,12 @@ export class Cli {
       });
     }
     catch (err: any) {
+      // restoring the command and logger is done here instead of in a 'finally' because there were issues with the code coverage tool
+      // restore the original command name
+      cli.currentCommandName = parentCommandName;
+      // restore the original logger
+      request.logger = currentLogger;
+
       throw {
         error: err,
         stderr: logErr.join(os.EOL)
@@ -395,21 +412,63 @@ export class Cli {
       alias: {}
     };
 
+    let argsToParse = args;
+
     if (commandInfo) {
       const commandTypes = commandInfo.command.types;
       if (commandTypes) {
         minimistOptions.string = commandTypes.string;
-        minimistOptions.boolean = commandTypes.boolean;
+
+        // minimist will parse unused boolean options to 'false' (unused options => options that are not included in the args) 
+        // But in the CLI booleans are nullable. They can can be true, false or undefined.
+        // For this reason we only pass boolean types that are actually used as arg.
+        minimistOptions.boolean = commandTypes.boolean.filter(optionName => args.some(arg => `--${optionName}` === arg || `-${optionName}` === arg));
       }
+
       minimistOptions.alias = {};
       commandInfo.options.forEach(option => {
         if (option.short && option.long) {
           (minimistOptions.alias as any)[option.short] = option.long;
         }
       });
+
+      argsToParse = this.getRewrittenArgs(args, commandTypes);
     }
 
-    return minimist(args, minimistOptions);
+    return minimist(argsToParse, minimistOptions);
+  }
+
+  /**
+   * Rewrites arguments (if necessary) before passing them into minimist.
+   * Currently only boolean values are checked and fixed.
+   * Args are only checked and rewritten if the option has been added to the 'types.boolean' array.
+   */
+  private getRewrittenArgs(args: string[], commandTypes: CommandTypes): string[] {
+    const booleanTypes = commandTypes.boolean;
+
+    if (booleanTypes.length === 0) {
+      return args;
+    }
+
+    return args.map((arg: string, index: number, array: string[]) => {
+      if (arg.startsWith('-') || index === 0) {
+        return arg;
+      }
+
+      // This line checks if the current arg is a value that belongs to a boolean option.
+      if (booleanTypes.some(t => `--${t}` === array[index - 1] || `-${t}` === array[index - 1])) {
+        const rewrittenBoolean = formatting.rewriteBooleanValue(arg);
+
+        if (!validation.isValidBoolean(rewrittenBoolean)) {
+          const optionName = array[index - 1];
+          throw new Error(`The value '${arg}' for option '${optionName}' is not a valid boolean`);
+        }
+
+        return rewrittenBoolean;
+      }
+
+      return arg;
+    });
   }
 
   private static formatOutput(logStatement: any, options: GlobalOptions): any {
@@ -576,12 +635,12 @@ export class Cli {
     return undefined;
   }
 
-  private printHelp(exitCode: number = 0): void {
+  private printHelp(helpMode: string, exitCode: number = 0): void {
     const properties: any = {};
 
     if (this.commandToExecute) {
       properties.command = this.commandToExecute.name;
-      this.printCommandHelp();
+      this.printCommandHelp(helpMode);
     }
     else {
       Cli.log();
@@ -602,7 +661,7 @@ export class Cli {
     process.exit(exitCode);
   }
 
-  private printCommandHelp(): void {
+  private printCommandHelp(helpMode: string): void {
     let helpFilePath = '';
     let commandNameWords: string[] = [];
     if (this.commandToExecute) {
@@ -625,9 +684,74 @@ export class Cli {
     helpFilePath = path.join(...pathChunks);
 
     if (fs.existsSync(helpFilePath)) {
+      let helpContents = fs.readFileSync(helpFilePath, 'utf8');
+      helpContents = this.getHelpSection(helpMode, helpContents);
+      helpContents = md.md2plain(helpContents, path.join(this.commandsFolder, '..', '..', 'docs'));
       Cli.log();
-      Cli.log(md.md2plain(fs.readFileSync(helpFilePath, 'utf8'), path.join(this.commandsFolder, '..', '..', 'docs')));
+      Cli.log(helpContents);
     }
+  }
+
+  private getHelpMode(options: any): string {
+    const h = options.h;
+    const help = options.help;
+
+    // user passed -h or --help, let's see if they passed a specific mode
+    // or requested the default
+    if (h || help) {
+      const helpMode: boolean | string = h ?? help;
+
+      if (typeof helpMode === 'boolean' || typeof helpMode !== 'string') {
+        // requested default mode or passed a number, let's use default
+        return this.getSettingWithDefaultValue<string>(settingsNames.helpMode, Cli.defaultHelpMode);
+      }
+      else {
+        const lowerCaseHelpMode = helpMode.toLowerCase();
+
+        if (Cli.helpModes.indexOf(lowerCaseHelpMode) < 0) {
+          Cli.getInstance().closeWithError(`Unknown help mode ${helpMode}. Allowed values are ${Cli.helpModes.join(', ')}`, { options }, false);
+          return ''; // noop
+        }
+        else {
+          return lowerCaseHelpMode;
+        }
+      }
+    }
+    else {
+      return this.getSettingWithDefaultValue<string>(settingsNames.helpMode, Cli.defaultHelpMode);
+    }
+  }
+
+  private getHelpSection(helpMode: string, helpContents: string): string {
+    if (helpMode === 'full') {
+      return helpContents;
+    }
+
+    // options is the first section, so get help up to options
+    const titleAndUsage = helpContents.substring(0, helpContents.indexOf('## Options'));
+
+    // find the requested section
+    const sectionLines: string[] = [];
+    const sectionName = helpMode[0].toUpperCase() + helpMode.substring(1);
+    const lines: string[] = helpContents.split('\n');
+    for (let i: number = 0; i < lines.length; i++) {
+      const line = lines[i];
+
+      if (line.indexOf(`## ${sectionName}`) === 0) {
+        sectionLines.push(line);
+      }
+      else if (sectionLines.length > 0) {
+        if (line.indexOf('## ') === 0) {
+          // we've reached the next section, stop
+          break;
+        }
+        else {
+          sectionLines.push(line);
+        }
+      }
+    }
+
+    return titleAndUsage + sectionLines.join('\n');
   }
 
   private printAvailableCommands(): void {
@@ -754,7 +878,7 @@ export class Cli {
 
     if (showHelpIfEnabled &&
       this.getSettingWithDefaultValue<boolean>(settingsNames.showHelpOnFailure, showHelpIfEnabled)) {
-      this.printHelp(exitCode);
+      this.printHelp(this.getHelpMode(args.options), exitCode);
     }
     else {
       process.exit(exitCode);