+++ src/m365/planner/commands/plan/plan-get.ts
@@ -93,7 +93,7 @@ class PlannerPlanGetCommand extends GraphCommand {
 
   #initOptionSets(): void {
     this.optionSets.push(
-      ['id', 'title']
+      { options: ['id', 'title'] }
     );
   }