+++ src/m365/spfx/commands/project/project-doctor.spec.ts
@@ -475,6 +475,14 @@ describe(commands.PROJECT_DOCTOR, () => {
     assert.strictEqual(findings.length, 0);
   });
 
+  it('e2e: shows correct number of findings for a valid 1.16.1 project', async () => {
+    sinon.stub(command as any, 'getProjectRoot').callsFake(_ => path.join(process.cwd(), 'src/m365/spfx/commands/project/test-projects/spfx-1161-webpart-react'));
+
+    await command.action(logger, { options: {} } as any);
+    const findings: FindingToReport[] = log[0];
+    assert.strictEqual(findings.length, 0);
+  });
+
   it('supports debug mode', () => {
     const options = command.options;
     let containsOption = false;