+++ src/m365/spo/commands/customaction/customaction-remove.ts
@@ -94,7 +94,7 @@ class SpoCustomActionRemoveCommand extends SpoCommand {
   }
 
   #initOptionSets(): void {
-    this.optionSets.push(['id', 'title']);
+    this.optionSets.push({ options: ['id', 'title'] });
   }
 
   public async commandAction(logger: Logger, args: CommandArgs): Promise<void> {