+++ src/m365/spo/commands/eventreceiver/eventreceiver-remove.spec.ts
@@ -7,6 +7,7 @@ import { CommandInfo } from '../../../../cli/CommandInfo';
 import { Logger } from '../../../../cli/Logger';
 import Command, { CommandError } from '../../../../Command';
 import request from '../../../../request';
+import { pid } from '../../../../utils/pid';
 import { formatting } from '../../../../utils/formatting';
 import { sinonUtil } from '../../../../utils/sinonUtil';
 import commands from '../../commands';
@@ -34,6 +35,7 @@ describe(commands.EVENTRECEIVER_REMOVE, () => {
   before(() => {
     sinon.stub(auth, 'restoreAuth').callsFake(() => Promise.resolve());
     sinon.stub(appInsights, 'trackEvent').callsFake(() => { });
+    sinon.stub(pid, 'getProcessName').callsFake(() => '');
     auth.service.connected = true;
     commandInfo = Cli.getCommandInfo(command);
   });
@@ -72,7 +74,8 @@ describe(commands.EVENTRECEIVER_REMOVE, () => {
   after(() => {
     sinonUtil.restore([
       appInsights.trackEvent,
-      auth.restoreAuth
+      auth.restoreAuth,
+      pid.getProcessName
     ]);
     auth.service.connected = false;
   });