+++ src/m365/spo/commands/folder/FolderProperties.ts
@@ -1,3 +1,5 @@
+import { RoleDefinition } from "../roledefinition/RoleDefinition";
+
 export interface FolderProperties {
   Exists: boolean;
   IsWOPIEnabled: boolean;
@@ -9,4 +11,16 @@ export interface FolderProperties {
   TimeLastModified: string;
   UniqueId: string;
   WelcomePage: string;
+  ListItemAllFields: ListItemAllFields;
+}
+export interface ListItemAllFields {
+  RoleAssignments: RoleAssignment[];
+}
+export interface RoleAssignment {
+  Member: Member;
+  RoleDefinitionBindings: RoleDefinition[];
+}
+export interface Member {
+  PrincipalType: number;
+  PrincipalTypeString: string;
 }